- moved UrlSource and its repository into packrat - Application activates Spring Data Repositories - configured dependency versions in properties
