+++ gradle.d/30-java.gradle
@@ -14,12 +14,11 @@ subprojects {
 
         dependencies {
             compile (
-                    'org.apache.commons:commons-lang3:3.4',
-                    'commons-io:commons-io:2.4',
+                    "org.apache.commons:commons-lang3:${version_commons_lang}",
+                    "commons-io:commons-io:${version_commons_io}",
             )
             testCompile (
                     project(':test-base'),
-                    'org.assertj:assertj-core:3.3.0',
             )
         }
