+++ gradle.properties
@@ -18,3 +18,14 @@ version_npm=3.5.3
 
 # mongo_plugin
 version_mongo_plugin=0.11.0
+
+#apache
+version_commons_lang=3.4
+version_commons_io=2.4
+version_httpclient=4.5.2
+
+#quartz
+version_quartz=2.2.2
+
+#assertj
+version_assertj=3.3.0
