+++ harvester/src/test/java/com/github/reflectoring/infiniboard/harvester/source/url/UrlSourceRetrieveJobTest.java
@@ -1,20 +1,6 @@
 package com.github.reflectoring.infiniboard.harvester.source.url;
 
-import com.github.reflectoring.infiniboard.harvester.HarvesterApplication;
-import com.github.reflectoring.infiniboard.test.categories.MongoIntegrationTests;
-import org.junit.Assert;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.context.ApplicationContext;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-import java.util.Date;
 
 public class UrlSourceRetrieveJobTest {
