+++ test-base/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
     compile(
-            'org.assertj:assertj-core:3.3.0'
+            "org.assertj:assertj-core:${version_assertj}",
     )
 }
